Code Review of HTML file for Dasmotos Brief
I have edited the html to reflect my suggestions on improving the code.
I deleted lines 7 & 8 about free stock images from canva as they aren't relevant to the brief.
I removed the header tags as the comment mentioned the nav was in a header tag in order to fix it to the top of the page when scrolling, but this can be achieved with ' nav { position: fixed;} ' in the css file.
I edited the indentation to conform with W3C standards, and put each <a> tag on a new line for easier readability. 
 
Originally on line 22, now on line 16, I changed the H1 tags to H2 for Tools as there should only be one H1 tag per page. I assumed that this means the Tools heading is above the over headings in hierarchy so I amended the remaining H2 and H3 tags to H3 and H4. 
As each Div has a class assigned, I changed Class PH, FH and BH to IDs since they aren't repeated on the page. 
I deleted the width and height attributes from the image that was originally on line 33 as it's better to specify this in CSS.