Code Review CSS
I've edited the code to reflect my suggestions.
I deleted the first two lines because only code inside the body tags is visible so styling should be applied to the Body tag rather than the HTML tag to apply a style to the whole page. You can also apply styling to '*' to style the whole page. I also removed the background colour because it was white and the page defaults to white anyway.

I reduced the top: 10px for the nav bar to 0px to avoid seeing the content underneath above the Nav bar as you scroll.
The default styling on anchor tags was overriding your styling on the text in the nav bar so I made all the styling such as text colour, weight etc specific to anchor tags as the ones in the navbar are the only anchor tags on the page. Otherwise I would've given the anchor tags in the navbar a class.

I removed the styling for the Shop class and the Brush1 class as the HTML file doesn't contain a Shop class or a Brush1 class.

I changed BH, FH, and PH to IDs to match my HTML edits.
The styling for BH wasn't visible because it had been coded to only apply to BH AND Brush1, but Brush1 doesn't exist. 
I changed the background colours for the headings to meet the brief.
I removed the image positioning as it's not part of the brief and the styling wasn't consistent across the page.